+++ lib/tmuxinator.rb
@@ -20,9 +20,3 @@ require "tmuxinator/version"
 
 module Tmuxinator
 end
-
-class Object
-  def blank?
-    respond_to?(:empty?) ? !!empty? : !self
-  end
-end