+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Remove unused attr_readers from Tmuxinator::Window
 - Add ability for pre_window commands to parse yaml arrays
+- Remove Object#blank? monkey patch (#458)
 
 ### Misc
 - Removed support for Ruby 1.9.3